refactor: constants 폴더 내 파일명을 소문자로 수정 및 해당 파일 내용 확장을 고려하여 s 추가
