+++ src/components/LunchPickerApp.js
@@ -1,7 +1,7 @@
 import Component from './Component';
 import RestaurantRepository from '../domain/RestaurantRepository';
 import { $, $setAttribute } from '../utils/dom';
-import { OPTION } from '../constants/Condition';
+import { OPTION } from '../constants/conditions';
 
 class LunchPickerApp extends Component {
   #restaurants;
