+++ src/constants/conditions.ts
